Good Testing Backup
Productive day of integration testing. A lot of items are now more fully functioning. Tons of work on the Galil interface. Testing with return to home, etc. Will add some of the stop code functionality in at a future date. Realized the sensoray does not have support really for modbus unless we are extremely clever about it.