Updates Communications
This commit handles some missing communication status messages. Also, this commit begins to address locking the command buttons when not in a state allowing the user to issue them. While this is not an issue with commands, changing values would be reflected in the GUI without actually having the command reflect it.