Logging Profile Variables
This commit should update the logging functionality to include profile variables. Upon a user chaging a variable on the GUI, it should upate the appropriate program profile within the state interface after receiving an acknowledgement of the command being completed. It currently does not handle changes that may be committed outside of the direct GUI. It would be ideal to determine a method to download the current galil variables before executing the script for logging.