Minimum Viable Automation
This commit handles the minimal viable automation when runnig a profile. The sequence of events will when hitting the run individual profile button shall turn on the pump, wait for the initialization time, setup any/all logging required, and then execute the profile.

Upon the profile being completed and/or aborted, the pump shall be turned off, logging closed, etc.