Pump and Manual Posiitoning
This commit corrects issues when operating with both the pump and manual positioning states.

The pump state had an issue if trying to execute a profile and would remain in the current state. This was caused by the conditional statements that stateud if pump was not initialized catching should not run. This was corrected.

Additionally, the galil had an issue with small relative motion commands and would remain in the state. This was caused by relatively small motion and with a significant speed. The inner state machine was originally only watching for stop codes. The speed at which the machine moved and settled was faster than that the code was polling. This would cause the state machine to never see the change and therefore remain in the state. This was remedied by also computing a position for the machine to target and additionally monitoring it. We should expose a threshold epsilon value in the settings. We also may have to consider increasing the polling rate.